add constant optimization for cosine similarity and normalization
